requirements.txt
keras==2.9.0
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
seaborn==0.11.2
tensorflow_gpu==2.9.0
